requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
langchain-openai
nltk
transformers

-e .